refactor : package 정리
- chess package 하위에 piece package가 존재합니다.
- chess package : Board, Color
- piece package : Pawn
- test 폴더도 함께 정리합니다.
