+++ main/chess/Color.java
@@ -1,4 +1,4 @@
-package piece;
+package chess;
 
 public enum Color {
     WHITE, BLACK
